refactor(cli): migrate caravan tooling onto discovery
Move caravan's Zenodo drift refresher (with its discovery/hydrology-filter
helpers) and structural validator to earthlens.caravan.cli, register them in the
ocean earthlens.cli table, and drop them from core's refresh/validate. Relocate
the caravan CLI tests to earthlens-ocean. Shrink the guard allow-lists (caravan;
caravan import).
